move the fake IrisAO class into pastis_functions
